feat(autoapi): export relocated endpoints
