+++ pkgs/standards/autoapi/autoapi/v2/endpoints/__init__.py
@@ -1,0 +1,6 @@
+"""Diagnostic and RPC endpoints for the AutoAPI framework."""
+
+__all__ = ["attach_health_and_methodz", "build_rpcdispatch"]
+
+from .endpoints import attach_health_and_methodz
+from .rpcdispatcher import build_rpcdispatch
