+++ pkgs/standards/autoapi/autoapi/v2/endpoints/rpcdispatcher.py
@@ -1,5 +1,5 @@
 """
-autoapi.v2.rpcdispatch  – JSON-RPC façade for AutoAPI
+autoapi.v2.endpoints.rpcdispatcher – JSON-RPC façade for AutoAPI
 """
 
 from __future__ import annotations
@@ -9,9 +9,9 @@ from typing import Any, Dict
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import Session
 
-from .types import APIRouter, Body, Depends, HTTPException, Request, ValidationError
-from .impl._runner import _invoke
-from .jsonrpc_models import (
+from ..types import APIRouter, Body, Depends, HTTPException, Request, ValidationError
+from ..impl._runner import _invoke
+from ..jsonrpc_models import (
     _RPCReq,
     _RPCRes,
     _err,
@@ -56,7 +56,9 @@ def build_rpcdispatch(api) -> APIRouter:
             except ValidationError as exc:
                 errors = exc.errors()
                 missing = [e["loc"][-1] for e in errors if e["type"] == "missing"]
-                msg = "Validation error" + (": missing parameter(s): " + ", ".join(missing) if missing else "")
+                msg = "Validation error" + (
+                    ": missing parameter(s): " + ", ".join(missing) if missing else ""
+                )
                 return _err(-32602, msg, env, data=errors)
             except Exception as exc:
                 return _err(-32000, str(exc), env)
@@ -93,7 +95,9 @@ def build_rpcdispatch(api) -> APIRouter:
             except ValidationError as exc:
                 errors = exc.errors()
                 missing = [e["loc"][-1] for e in errors if e["type"] == "missing"]
-                msg = "Validation error" + (": missing parameter(s): " + ", ".join(missing) if missing else "")
+                msg = "Validation error" + (
+                    ": missing parameter(s): " + ", ".join(missing) if missing else ""
+                )
                 return _err(-32602, msg, env, data=errors)
             except Exception as exc:
                 return _err(-32000, str(exc), env)
